WIP refactored side panel into own component
