+++ apps/frontend/src/app/backups-overview/component/backups.component.ts
@@ -1,287 +1,17 @@
-import { AfterViewInit, Component, OnDestroy, OnInit } from '@angular/core';
-import {
-  BehaviorSubject,
-  combineLatest,
-  debounceTime,
-  distinctUntilChanged,
-  map,
-  Observable,
-  shareReplay,
-  startWith,
-  Subject,
-  switchMap,
-  takeUntil,
-  tap,
-} from 'rxjs';
-import { BackupService } from '../service/backup-service/backup-service.service';
-import { Backup } from '../../shared/types/backup';
-import { ClrDatagridSortOrder } from '@clr/angular';
-import { ChartService } from '../service/chart-service/chart-service.service';
-import { APIResponse } from '../../shared/types/api-response';
-import { BackupTask } from '../../shared/types/backup.task';
-import { BackupType } from '../../shared/enums/backup.types';
-
-interface TimeRangeConfig {
-  fromDate: Date;
-  toDate: Date;
-  range: 'week' | 'month' | 'year';
-}
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-backups',
   templateUrl: './backups.component.html',
   styleUrl: './backups.component.css',
 })
-export class BackupsComponent implements AfterViewInit, OnDestroy, OnInit {
-  protected readonly ClrDatagridSortOrder = ClrDatagridSortOrder;
-
-  protected timeRanges: ('week' | 'month' | 'year')[] = [
-    'week',
-    'month',
-    'year',
-  ];
-  tasksLoading = false;
-  loading = false;
-  pageSize = 10;
-
-  backupEnumTypes = Object.keys(BackupType).filter((item) => {
-    return isNaN(Number(item));
-  });
-
-  //Filters for Charts
-  selectedBackupTypesForCharts: string[] = [];
-  protected selectedTask: BackupTask[] = [];
+export class BackupsComponent {
   protected filterPanel = false;
-
-  //Subjects
-  private readonly timeRangeSubject$ = new BehaviorSubject<TimeRangeConfig>({
-    fromDate: new Date(),
-    toDate: new Date(),
-    range: 'month',
-  });
-  readonly timeRange$ = this.timeRangeSubject$.pipe(
-    map((config) => config.range)
-  );
-
-  protected backupTaskSearchTerm$: Subject<string> = new Subject<string>();
-
-  readonly backupTaskSubject$ = new BehaviorSubject<BackupTask[]>([]);
-  readonly backupTypesForChartsSubject$ = new BehaviorSubject<BackupType[]>([]);
-
-  private readonly destroy$ = new Subject<void>();
-
-  //Observables
-  readonly chartBackups$: Observable<APIResponse<Backup>>;
-  allBackupTasks$: Observable<BackupTask[]>;
-  protected selectedbackupTasks$: Observable<BackupTask[]>;
-
-  constructor(
-    private readonly backupService: BackupService,
-    private readonly chartService: ChartService
-  ) {
-    /**
-     * Load all backups and filter them based on the filter options for charts
-     */
-    this.allBackupTasks$ = this.backupService
-      .getAllBackupTasks()
-      .pipe(takeUntil(this.destroy$), shareReplay(1));
-
-    this.selectedbackupTasks$ = combineLatest([
-      this.allBackupTasks$,
-      this.backupTaskSearchTerm$.pipe(
-        startWith(''),
-        debounceTime(300),
-        distinctUntilChanged()
-      ),
-    ]).pipe(
-      map(([tasks, searchTerm]) => {
-        if (!searchTerm) {
-          return [];
-        }
-
-        const term = searchTerm.toLowerCase();
-        return tasks.filter((task) =>
-          task.displayName.toLowerCase().includes(term)
-        );
-      })
-    );
-
-    this.chartBackups$ = combineLatest([
-      this.timeRangeSubject$.pipe(
-        distinctUntilChanged(
-          (prev, curr) =>
-            prev.range === curr.range &&
-            prev.fromDate.getTime() === curr.fromDate.getTime() &&
-            prev.toDate.getTime() === curr.toDate.getTime()
-        )
-      ),
-      this.backupTaskSubject$.pipe(
-        distinctUntilChanged((prev, curr) => {
-          if (!prev && !curr) return true;
-          if (!prev || !curr) return false;
-          if (prev.length !== curr.length) {
-            return false;
-          }
-          const prevIds = prev.map((p) => p.id).sort();
-          const currIds = curr.map((c) => c.id).sort();
-
-          return prevIds === currIds;
-        })
-      ),
-      this.backupTypesForChartsSubject$.pipe(
-        distinctUntilChanged((prev, curr) => {
-          if (!prev && !curr) return true;
-          if (!prev || !curr) return false;
-          if (prev.length !== curr.length) {
-            return false;
-          }
-          const prevIds = prev.map((p) => p).sort();
-          const currIds = curr.map((c) => c).sort();
-          return prevIds === currIds;
-        })
-      ),
-    ]).pipe(
-      map(([timeRange, tasks, backupTypes]) => ({
-        params: {
-          fromDate: timeRange.fromDate.toISOString(),
-          toDate: timeRange.toDate.toISOString(),
-          types: backupTypes,
-        },
-        selectedTasks: tasks ? tasks.map((task) => task.id) : [],
-      })),
-      switchMap(({ params, selectedTasks }) =>
-        this.backupService.getAllBackups(params, selectedTasks)
-      ),
-      tap({
-        next: (response) => {
-          if (response.data && response.data.length > 0) {
-            const currentRange = this.timeRangeSubject$.getValue().range;
-            // Update timeline chart
-            const columnData = this.chartService.prepareColumnData(
-              response.data,
-              currentRange
-            );
-            this.chartService.updateChart('backupTimelineChart', columnData);
-            // Update size distribution chart
-            const pieData = this.chartService.preparePieData(response.data);
-            this.chartService.updateChart('backupSizeChart', pieData);
-          } else {
-            console.warn('No data received for charts');
-          }
-          this.loading = false;
-        },
-        error: (error) => {
-          console.error('Error updating charts:', error);
-          this.loading = false;
-        },
-      }),
-      takeUntil(this.destroy$)
-    );
-  }
-
-  ngOnInit(): void {
-    this.setTimeRange('month');
-  }
-
-  /**
-   * Initialize the charts
-   */
-  ngAfterViewInit(): void {
-    setTimeout(() => {
-      this.chartService.createChart(
-        {
-          id: 'backupSizeChart',
-          type: 'pie',
-          valueField: 'value',
-          categoryField: 'category',
-          seriesName: 'SizeDistribution',
-        },
-        this.chartBackups$.pipe(
-          map((response: APIResponse<Backup>) => response.data)
-        )
-      );
-      this.chartService.createChart(
-        {
-          id: 'backupTimelineChart',
-          type: 'column',
-          valueYField: 'sizeMB',
-          valueXField: 'creationDate',
-          seriesName: 'BackupSize',
-          tooltipText:
-            "[bold]{valueY}[/] MB\n{valueX.formatDate('yyyy-MM-dd HH:mm')}\nBackups: {count}",
-        },
-        this.chartBackups$.pipe(
-          map((response: APIResponse<Backup>) => response.data)
-        ),
-        this.timeRangeSubject$.getValue().range
-      );
-    }, 100);
-  }
-
-  ngOnDestroy(): void {
-    this.destroy$.next();
-    this.destroy$.complete();
-    this.chartService.dispose();
-  }
-
-  /**
-   * Set time range for the charts
-   * @param range time range for the charts
-   */
-  setTimeRange(range: 'week' | 'month' | 'year'): void {
-    const toDate = new Date();
-    const fromDate = new Date();
-
-    switch (range) {
-      case 'week':
-        fromDate.setDate(fromDate.getDate() - 7);
-        break;
-      case 'month':
-        fromDate.setMonth(fromDate.getMonth() - 1);
-        break;
-      case 'year':
-        fromDate.setFullYear(fromDate.getFullYear() - 1);
-        break;
-    }
-
-    this.timeRangeSubject$.next({
-      fromDate,
-      toDate,
-      range,
-    });
-    this.chartService.updateTimeRange('backupTimelineChart', range);
-  }
-
-  /**
-   * Set selected Backup task to filter the charts
-   * @param tasks selected Backup task
-   */
-  setBackupTask(tasks: BackupTask[]): void {
-    this.selectedTask = tasks;
-    this.backupTaskSubject$.next(tasks);
-  }
-
-  setBackupTypesForCharts(types: BackupType[]): void {
-    this.selectedBackupTypesForCharts = types;
-    this.backupTypesForChartsSubject$.next(types);
-  }
-
-  /**
-   * Add search Term to backupTaskSearchTerm$ subject for the Backup task search
-   * @param term Search term for the Backup task
-   */
-  onSearchInput(term: string): void {
-    this.backupTaskSearchTerm$.next(term);
-  }
 
   /**
    * Change the state of the filter panel to open or close it
    */
   protected changeFilterPanelState(): void {
-    if (this.filterPanel) {
-      this.filterPanel = false;
-    } else {
-      this.filterPanel = true;
-    }
+    this.filterPanel = !this.filterPanel;
   }
 }
